major refactor and code cleanup part1

+++ utils/helpers.go
@@ -94,13 +94,13 @@ func ReadMultipartJSON(w http.ResponseWriter, r *http.Request, source interface{
 	return file, file_data, nil
 }
 
-func ReadIDParam(r *http.Request) (int64, error) {
+func ReadIDParam(r *http.Request) (int, error) {
 	params := httprouter.ParamsFromContext(r.Context())
 	id, err := strconv.ParseInt(params.ByName("id"), 10, 64)
 	if err != nil || id < 1 {
 		return 0, errors.New("invalid id parameter")
 	}
-	return id, nil
+	return int(id), nil
 }
 
 func ReadStringParam(qs url.Values, key string, defaultValue string) string {
